refactor: new compositor nodes structure
