+++ __init__.py
@@ -42,10 +42,10 @@ classes = [NodeToPythonMenu,
             options.NTPOptions,
             options.NTPOptionsPanel,
             #compositor
-            compositor.NTPCompositorOperator,
-            compositor.NTPCompositorScenesMenu,
-            compositor.NTPCompositorGroupsMenu,
-            compositor.NTPCompositorPanel,
+            compositor.operator.NTPCompositorOperator,
+            compositor.ui.NTPCompositorScenesMenu,
+            compositor.ui.NTPCompositorGroupsMenu,
+            compositor.ui.NTPCompositorPanel,
             #geometry
             geometry.operator.NTPGeoNodesOperator,
             geometry.menu.NTPGeoNodesMenu,
