+++ compositor/__init__.py
@@ -1,0 +1,13 @@
+if "bpy" in locals():
+    import importlib
+    importlib.reload(node_settings)
+    importlib.reload(node_tree)
+    importlib.reload(operator)
+    importlib.reload(ui)
+else:
+    from . import node_settings
+    from . import node_tree
+    from . import operator
+    from . import ui
+
+import bpy
